Fixed title screen movement speed glitch
For some reason Vector3.Lerp was causing the speed to exponentially rise, changing this to Vector3.Movetowards fixed this. However, the speed is currently relative to the amount of distance the camera has to travel for consistency's sake, though I'm not sure whether making it the same speed, or letting the time increase would be better. Alternatively, the canvases could be in the same place and just enabled/disabled, or to the left/right/above of the main canvas.